Integrated the model with app. The upload button responds with result of image identified.
