+++ templates/inspector.html
@@ -69,13 +69,16 @@
 
 
         </div>
-        <div id="results" class="col-md-6  hidden">
+        <div id="results" class="col-md-6">
           <h3>Results</h3>
-            <div id="heartdisease" class="hidden">
-                <strong>There is a probability of a heart disease.</strong>
+            <!-- <div id="heartdisease" class="hidden">
+                <strong>The image is identified as {result}</strong>
             </div>
             <div id="noheartdisease" class="hidden">
-                <strong>There is no probability of a heart disease.</strong>
+                <strong>Could not identify the image</strong>
+            </div> -->
+            <div id="foodresult">
+                <strong>The image is identified as {result}</strong>
             </div>
         </div>
       </div> <!-- end of div class row -->
